JSON update & Pathing Fix
Updated the JSON file to reflect an array of players and coaches within the users object. This means that each user will not have to contain a 'type' and can now just come under the umbrella of their type.

This also means that in the code we have to reference usersObj.users[0].players for players rather than usersObj.users.players because .users is now also an array and therefore requires the selector before one can access the object.

Added coach input validation which only runs after users have been checked, this has been done because i am expecting the players array to be much larger than the coaches array, so statistically the code should be run less.