Storing current user details
Created a getter and a setter in the flask script for the current user details.

Now user details are accessible from any point within the system, this is especially helpful for the user and the coach dashboard pages where we will be able to dynamically load names and information rather than have "name" "name" etc.